Move to new config.e2e.ts
